+++ src/model/index.ts
@@ -1,4 +1,5 @@
 export * from "./build";
 export * from "./lib";
+export * from "./tags";
 export * from "./types";
 export * from "./validate";